requirements.txt
asgiref==3.11.0
channels==4.3.2
Django==6.0.1
django-cors-headers==4.9.0
django-extensions==4.1
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Markdown==3.10
pillow==12.1.0
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
